% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_complaints.R
\name{check_response_status}
\alias{check_response_status}
\title{Check the API response to be sure it's a-okay}
\usage{
check_response_status(query_results, cfpb_query_path)
}
\arguments{
\item{query_results}{(string) A list of arguments to be passed in the query URL}
}
\description{
Check the API response to be sure it's a-okay
}
