% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_complaints.R
\name{get_all_complaints}
\alias{get_all_complaints}
\title{Downloads and reads a zip file containing all public complaints.}
\usage{
get_all_complaints(quiet = TRUE)
}
\arguments{
\item{quiet}{Should progress messages be printed?}
}
\value{
A data frame:
\describe{
\item{\code{product}}{factor The type of product the consumer identified in the complaint}
\item{\code{complaint_what_happened}}{string Consumer complaint narrative is the consumer-submitted description of "what happened" from the complaint.}
\item{\code{date_sent_to_company}}{POSIXlt The date the CFPB sent the complaint to the company}
\item{\code{issue}}{factor The issue the consumer identified in the complaint}
\item{\code{sub_product}}{factor The type of sub-product the consumer identified in the complaint}
\item{\code{zip_code}}{string}
\item{\code{tags}}{Data that supports easier searching and sorting of complaints submitted by or on behalf of consumers.}
\item{\code{has_narrative}}{boolean}
\item{\code{complaint_id}}{integer }
\item{\code{timely}}{factor Whether the company gave a timely response}
\item{\code{consumer_consent_provided}}{factor Identifies whether the consumer opted in to publish their complaint narrative.}
\item{\code{company_response}}{string }
\item{\code{submitted_via}}{factor How the complaint was submitted to the CFPB}
\item{\code{company}}{string The complaint is about this company}
\item{\code{date_received}}{POSIXlt The date the CFPB received the complaint}
\item{\code{state}}{factor}
\item{\code{consumer_disputed}}{factor Whether the consumer disputed the company’s response}
\item{\code{company_public_response}}{string The company's optional, public-facing response to a consumer's complaint}
\item{\code{sub_issue}}{factor The sub-issue the consumer identified in the complaint}
}
}
\description{
Downloads and reads a zip file containing all public complaints.
}
\examples{
\dontrun{
  test <- get_all_complaints()
}
}
