% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_cfpb_url_path}
\alias{get_cfpb_url_path}
\title{Get CFPB URL path}
\usage{
get_cfpb_url_path(pathend)
}
\arguments{
\item{pathend}{the string or string vector of ending url paths}
}
\value{
CFPB URL path
}
\description{
Get CFPB URL path
}
