% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match.call.defaults}
\alias{match.call.defaults}
\title{Get function arguments, including defaults}
\usage{
match.call.defaults(...)
}
\value{
arguments
}
\description{
Get function arguments, including defaults
}
