% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_companies.R
\name{query_companies}
\alias{query_companies}
\title{Suggest Possible Companies based on partial match}
\usage{
query_companies(text, state = NULL, zip_code = NULL)
}
\arguments{
\item{text}{required (string) text to use for suggestions}

\item{state}{(string array) Filter the results to only return results in these states (use abbreviation, i.e. CA, VA)}

\item{zip_code}{(string array) Filter the results to only return results in a given zip code}
}
\value{
character vector matching text string
}
\description{
API docs:  https://cfpb.github.io/api/ccdb/
}
\examples{
\dontrun{
  query_companies(text = "chase")
}
}
