% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_complaints.R
\name{query_complaints}
\alias{query_complaints}
\title{Query For Complaints}
\usage{
query_complaints(
  search_term = NULL,
  field = "complaint_what_happened",
  size = 10000,
  company = NULL,
  company_public_response = NULL,
  company_received_max = NULL,
  company_received_min = NULL,
  company_response = NULL,
  consumer_consent_provided = NULL,
  consumer_disputed = NULL,
  date_received_max = NULL,
  date_received_min = NULL,
  has_narrative = NULL,
  issue = NULL,
  product = NULL,
  state = NULL,
  submitted_via = NULL,
  tags = NULL,
  timely = NULL,
  zip_code = NULL
)
}
\arguments{
\item{search_term}{(string) Return results containing specific term}

\item{field}{(string) If the parameter "search_term" has a value, use "field" to specify which field is searched. If not specified, "complaint_what_happened" will be searched.}

\item{size}{(integer) Limit the size of the results.}

\item{company}{(string array) Filter the results to only return these companies}

\item{company_public_response}{(string array) Filter the results to only return these types of public response by the company}

\item{company_received_max}{(string) Return results with date < company_received_max (i.e. 2017-03-04)}

\item{company_received_min}{(string) Return results with date >= company_received_min (i.e. 2017-03-04)}

\item{company_response}{(string array) Filter the results to only return these types of response by the company. Available strings are 'Closed', 'Closed with explanation', 'Closed with monetary relief', 'Closed with non-monetary relief', 'Closed with relief', 'Closed without relief', 'In progress', or 'Untimely response'.}

\item{consumer_consent_provided}{(string array) Filter the results to only return these types of consent consumer provided. Available strings are 'Consent not provided', 'Consent provided', 'Consent withdrawn', 'N/A', or 'Other'.}

\item{consumer_disputed}{(string array) Filter the results to only return the specified state of consumer disputed. Available values are 'Yes', 'No', or 'N/A'.}

\item{date_received_max}{(string) Return results with date < date_received_max (i.e. 2017-03-04)}

\item{date_received_min}{(string) Return results with date >= date_received_min (i.e. 2017-03-04)}

\item{has_narrative}{(string array) Filter the results to only return the specified state of whether it has narrative in the complaint or not. Available values are 'true', or 'false'.}

\item{issue}{(string array) Filter the results to only return these types of issue and subissue, i.e. product-only: Getting a Loan, subproduct needs to include product, separated by '•', Getting a Loan•Cant qualify for a loan}

\item{product}{(string array) Filter the results to only return these types of product and subproduct, i.e. product-only: Mortgage, subproduct needs to include product, separated by '•', Mortgage•FHA mortgage}

\item{state}{(string array) Filter the results to only return these states (use abbreviation, i.e. CA, VA)}

\item{submitted_via}{(string array) Filter the results to only return these types of way consumers submitted their complaints. Available values are 'Fax', 'Phone', 'Postal mail', 'Referral', 'Web', or 'Web Referral'.}

\item{tags}{(string array) Filter the results to only return these types of tag. Available tags are 'Older American', 'Servicemember', and 'Older American, Servicemember'.}

\item{timely}{(string array) Filter the results to show whether a response was timely. Available values are 'Yes', or 'No'.}
}
\value{
A data frame:
\describe{
\item{\code{product}}{factor The type of product the consumer identified in the complaint}
\item{\code{complaint_what_happened}}{string Consumer complaint narrative is the consumer-submitted description of "what happened" from the complaint.}
\item{\code{date_sent_to_company}}{POSIXlt The date the CFPB sent the complaint to the company}
\item{\code{issue}}{factor The issue the consumer identified in the complaint}
\item{\code{sub_product}}{factor The type of sub-product the consumer identified in the complaint}
\item{\code{zip_code}}{string}
\item{\code{tags}}{Data that supports easier searching and sorting of complaints submitted by or on behalf of consumers.}
\item{\code{has_narrative}}{boolean}
\item{\code{complaint_id}}{integer }
\item{\code{timely}}{factor Whether the company gave a timely response}
\item{\code{consumer_consent_provided}}{factor Identifies whether the consumer opted in to publish their complaint narrative.}
\item{\code{company_response}}{string }
\item{\code{submitted_via}}{factor How the complaint was submitted to the CFPB}
\item{\code{company}}{string The complaint is about this company}
\item{\code{date_received}}{POSIXlt The date the CFPB received the complaint}
\item{\code{state}}{factor}
\item{\code{consumer_disputed}}{factor Whether the consumer disputed the company’s response}
\item{\code{company_public_response}}{string The company's optional, public-facing response to a consumer's complaint}
\item{\code{sub_issue}}{factor The sub-issue the consumer identified in the complaint}
}
}
\description{
API docs:  https://cfpb.github.io/api/ccdb/
}
\examples{
\dontrun{
  query_complaints(size = 1)
}

}
