% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_complaints.R
\name{query_no_page}
\alias{query_no_page}
\title{Query without paging}
\usage{
query_no_page(cfpb_query_list)
}
\arguments{
\item{cfpb_query_list}{(list) A list of arguments to be passed in the query URL}
}
\description{
Query without paging
}
