% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_complaints.R
\name{query_page}
\alias{query_page}
\title{Query with paging}
\usage{
query_page(cfpb_query_list)
}
\arguments{
\item{cfpb_query_list}{(list) A list of arguments to be passed in the query URL}
}
\description{
Query with paging
}
