% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_aggregates.R
\name{query_products}
\alias{query_products}
\title{Query for complaint information at product level}
\usage{
query_products(
  search_term = NULL,
  field = "complaint_what_happened",
  company = NULL,
  company_public_response = NULL,
  company_received_max = NULL,
  company_received_min = NULL,
  company_response = NULL,
  consumer_consent_provided = NULL,
  consumer_disputed = NULL,
  date_received_max = NULL,
  date_received_min = NULL,
  has_narrative = NULL,
  issue = NULL,
  state = NULL,
  submitted_via = NULL,
  tags = NULL,
  timely = NULL
)
}
\arguments{
\item{search_term}{(string) Return results containing specific term}

\item{field}{(string) If the parameter "search_term" has a value, use "field" to specify which field is searched. If not specified, "complaint_what_happened" will be searched.}

\item{company}{(string array) Filter the results to only return these companies}

\item{company_public_response}{(string array) Filter the results to only return these types of public response by the company}

\item{company_received_max}{(string) Return results with date < company_received_max (i.e. 2017-03-04)}

\item{company_received_min}{(string) Return results with date >= company_received_min (i.e. 2017-03-04)}

\item{company_response}{(string array) Filter the results to only return these types of response by the company}

\item{consumer_consent_provided}{(string array) Filter the results to only return these types of consent consumer provided}

\item{consumer_disputed}{(string array) Filter the results to only return the specified state of consumer disputed, i.e. yes, no}

\item{date_received_max}{(string) Return results with date < date_received_max (i.e. 2017-03-04)}

\item{date_received_min}{(string) Return results with date >= date_received_min (i.e. 2017-03-04)}

\item{has_narrative}{(string array) Filter the results to only return the specified state of whether it has narrative in the complaint or not, i.e. yes, no}

\item{issue}{(string array) Filter the results to only return these types of issue and subissue, i.e. product-only: Getting a Loan, subproduct needs to include product, separated by '•', Getting a Loan•Cant qualify for a loan}

\item{state}{(string array) Filter the results to only return these states (use abbreviation, i.e. CA, VA)}

\item{submitted_via}{(string array) Filter the results to only return these types of way consumers submitted their complaints}

\item{tags}{(string array) Filter the results to only return these types of tag}

\item{timely}{(string array) Filter the results to show whether a response was timely}
}
\value{
A data frame:
\describe{
\item{\code{product}}{string}
\item{\code{subproduct}}{string}
\item{\code{doc_count}}{Number of complaints returned from the search}
}
}
\description{
API docs:  https://cfpb.github.io/api/ccdb/
}
\examples{
\dontrun{
  query_complaints(size = 1)
}

}
