% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_trends.R
\name{query_trends}
\alias{query_trends}
\title{Query for complaint trends}
\usage{
query_trends(
  search_term = NULL,
  lens = "overview",
  trend_interval = "year",
  field = "complaint_what_happened",
  company = NULL,
  company_public_response = NULL,
  company_received_max = NULL,
  company_received_min = NULL,
  company_response = NULL,
  consumer_consent_provided = NULL,
  consumer_disputed = NULL,
  date_received_max = NULL,
  date_received_min = NULL,
  has_narrative = NULL,
  issue = NULL,
  product = NULL,
  state = NULL,
  submitted_via = NULL,
  tags = NULL,
  timely = NULL,
  trend_depth = 10,
  zip_code = NULL
)
}
\arguments{
\item{search_term}{(string) Return results containing specific term}

\item{lens}{(string) The data lens through which to view complaint trends over time. Available values are overview (default), issue, product, or tags.}

\item{trend_interval}{(string) The interval of time to use for aggregations. When using day intervals, we recommend querying for periods of less than one year. Available values: year (default), quarter, month, week, day}

\item{field}{(string) If the parameter "search_term" has a value, use "field" to specify which field is searched. Available values: complaint_what_happened (default), company_public_response, all.}

\item{company}{(string array) Filter the results to only return these companies}

\item{company_public_response}{(string array) Filter the results to only return these types of public response by the company}

\item{company_received_max}{(string) Return results with date < company_received_max (i.e. 2017-03-04)}

\item{company_received_min}{(string) Return results with date >= company_received_min (i.e. 2017-03-04)}

\item{company_response}{(string array) Filter the results to only return these types of response by the company}

\item{consumer_consent_provided}{(string array) Filter the results to only return these types of consent consumer provided}

\item{consumer_disputed}{(string array) Filter the results to only return the specified state of consumer disputed, i.e. yes, no}

\item{date_received_max}{(string) Return results with date < date_received_max (i.e. 2017-03-04)}

\item{date_received_min}{(string) Return results with date >= date_received_min (i.e. 2017-03-04)}

\item{has_narrative}{(string array) Filter the results to only return the specified state of whether it has narrative in the complaint or not, i.e. yes, no}

\item{issue}{(string array) Filter the results to only return these types of issue and subissue, i.e. product-only: Getting a Loan, subproduct needs to include product, separated by '•', Getting a Loan•Cant qualify for a loan}

\item{product}{(string array) Filter the results to only return these types of product and subproduct, i.e. product-only: Mortgage, subproduct needs to include product, separated by '•', Mortgage•FHA mortgage}

\item{state}{(string array) Filter the results to only return these states (use abbreviation, i.e. CA, VA)}

\item{submitted_via}{(string array) Filter the results to only return these types of way consumers submitted their complaints}

\item{tags}{(string array) Filter the results to only return these types of tag}

\item{timely}{(string array) Filter the results to show whether a response was timely}

\item{trend_depth}{(integer) The top X trend aggregations will be returned, where X is the supplied trend_depth. Defaults to 10}

\item{zip_code}{(string array) Zip Code}
}
\value{
A data frame containing the lens categories (e.g. product/issue/tag categories) and:
\describe{
\item{\code{timestamp}}{string}
\item{\code{doc_count}}{Number of complaints returned from the search}
}
}
\description{
API docs:  https://cfpb.github.io/api/ccdb/
}
\examples{
\dontrun{
  query_complaints(size = 1)
}

}
