% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_complaints.R
\name{to_dataframe}
\alias{to_dataframe}
\title{Construct dataframe from text-ified response}
\usage{
to_dataframe(text_res)
}
\arguments{
\item{text_res}{(string) The content of the API response, converted using jsonlite::FromJSON}
}
\description{
Construct dataframe from text-ified response
}
